clusters not being copied correctly

+++ apps/kmeans/kernel/kmeans.c
@@ -311,7 +311,7 @@ kmeans_result kmeans( const int64_t *points,  int64_t *centers,  int64_t *cluste
 	int iterations = 0;
     int max_iteration=20;
     
-    size_t clusters_sz = NUM_POINTS * sizeof(int);
+    size_t clusters_sz = NUM_POINTS * sizeof(int64_t);
 	
 	while (1)
 	{
